+++ include/gnfs/sqrt/algebraic_sqrt.hpp
@@ -318,7 +318,8 @@ private:
         for (int i = static_cast<int>(d) - 1; i >= 1; --i) {
             result *= m;
             // mpz_addmul_ui: result += nf.coeff(i) * i (i ≥ 1, fused FMA)
-            mpz_addmul_ui(result.get_mpz(), nf.coeff(i).get_mpz(), static_cast<unsigned long>(i));
+            mpz_addmul_ui(result.get_mpz(), nf.coeff(static_cast<uint32_t>(i)).get_mpz(),
+                          static_cast<unsigned long>(i));
             result %= N;
         }
         if (result.is_negative()) result += N;
